# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/reviews": {
      "post": {
        "operationId": "AlojamientosController_getReviews",
        "summary": "Obtener reseñas de alojamientos",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewsRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reseñas de los alojamientos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewsResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Búsqueda y Catálogo"
        ]
      }
    },
...
    "/api/v1/host/accommodations": {
      "get": {
        "operationId": "HostController_lista",
        "summary": "Mis alojamientos",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catálogo (ADMIN)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "HostController_crear",
        "summary": "Crear alojamiento (queda en BORRADOR)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearAlojamientoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Catálogo (ADMIN)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/host/accommodations/{codigo}": {
      "get": {
        "operationId": "HostController_obtener",
        "summary": "Detalle editable de mi alojamiento",
        "parameters": [
          {
            "name": "codigo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catálogo (ADMIN)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "HostController_actualizar",
        "summary": "Editar mi alojamiento",
        "parameters": [
          {
            "name": "codigo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarAlojamientoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catálogo (ADMIN)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/host/accommodations/{codigo}/publish": {
      "post": {
        "operationId": "HostController_publicar",
        "summary": "Publicar (requiere unidad activa y foto)",
        "parameters": [
          {
            "name": "codigo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catálogo (ADMIN)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/host/accommodations/{codigo}/unpublish": {
      "post": {
        "operationId": "HostController_despublicar",
        "summary": "Despublicar (vuelve a BORRADOR)",
        "parameters": [
          {
            "name": "codigo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catálogo (ADMIN)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/host/accommodations/{codigo}/units": {
      "get": {
        "operationId": "HostController_unidades",
        "summary": "Unidades (habitaciones/tipos) de mi alojamiento",
        "parameters": [
          {
            "name": "codigo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catálogo (ADMIN)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "HostController_crearUnidad",
        "summary": "Agregar unidad",
        "parameters": [
          {
            "name": "codigo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnidadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Catálogo (ADMIN)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/host/units/{id}": {
      "patch": {
        "operationId": "HostController_actualizarUnidad",
        "summary": "Editar o desactivar una unidad",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarUnidadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catálogo (ADMIN)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/host/units/{id}/calendar": {
      "get": {
        "operationId": "HostController_calendario",
        "summary": "Calendario de precio y cupo (máx. 93 días)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "desde",
            "required": true,
            "in": "query",
            "schema": {
              "pattern": "FECHA",
              "type": "string"
            }
          },
          {
            "name": "hasta",
            "required": true,
            "in": "query",
            "schema": {
              "pattern": "FECHA",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catálogo (ADMIN)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "HostController_guardarCalendario",
        "summary": "Guardar precios especiales / cupo por día",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarioDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catálogo (ADMIN)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/host/orders": {
      "get": {
        "operationId": "HostController_reservas",
        "summary": "Reservas y solicitudes de mis alojamientos",
        "parameters": [
          {
            "name": "estado",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PENDIENTE",
                "CONFIRMADA",
                "CANCELADA",
                "RECHAZADA",
                "EXPIRADA",
                "COMPLETADA"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catálogo (ADMIN)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/host/orders/{id}/respond": {
      "post": {
        "operationId": "HostController_responder",
        "summary": "Aceptar o rechazar una solicitud (fn_responder_solicitud)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponderSolicitudDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catálogo (ADMIN)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/host/income": {
      "get": {
        "operationId": "HostController_ingresos",
        "summary": "Ingresos por mes (v_ingresos_anfitrion)",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catálogo (ADMIN)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/host/reviews": {
      "get": {
        "operationId": "HostController_resenas",
        "summary": "Reseñas de mis alojamientos",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catálogo (ADMIN)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/host/reviews/{id}/reply": {
      "post": {
        "operationId": "HostController_responderResena",
        "summary": "Responder una reseña",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponderResenaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Catálogo (ADMIN)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Posada EC · Alojamientos API",
    "description": "Implementación del contrato \"GDS Alojamientos Core API\" (rutas del YAML bajo /api/v1) y rutas propias (auth, cuenta, anfitrión, admin). El ownerId de cada reserva se infiere del `sub` del JWT. OAuth2 simplificado: el token se obtiene con POST /api/v1/auth/login y se envía como Bearer.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Este servidor"
    }
  ],
  "components": {
    "securitySchemes": {
      "OAuth2Security": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://auth.booking-hub.com/oauth2/authorize",
            "tokenUrl": "https://auth.booking-hub.com/oauth2/token",
            "scopes": {
              "alojamientos:read": "Leer información de alojamientos, disponibilidades y reservas",
              "alojamientos:book": "Crear y alterar reservas",
              "alojamientos:cancel": "Cancelar reservas",
              "alojamientos:webhooks": "Gestionar webhooks"
            }
          },
          "clientCredentials": {
            "tokenUrl": "https://auth.booking-hub.com/oauth2/token",
            "scopes": {
              "alojamientos:read": "Leer información de alojamientos, disponibilidades y reservas",
              "alojamientos:book": "Crear y alterar reservas",
              "alojamientos:cancel": "Cancelar reservas",
              "alojamientos:webhooks": "Gestionar webhooks"
            }
          }
        },
        "description": "OAuth2 del contrato. En este prototipo el token Bearer se emite en /api/v1/auth/login."
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ReviewsRequestDto": {
        "type": "object",
        "properties": {
          "accommodations": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "page": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "rows": {
            "type": "number",
            "minimum": 1,
            "maximum": 100
          },
          "languages": {
            "description": "Solo \"es\"",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "accommodations"
        ]
      },
      "Resena": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string",
            "nullable": true
          },
          "host_reply": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "accommodation_id": {
            "type": "number"
          },
          "score": {
            "type": "number"
          },
          "author": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "comment",
          "host_reply",
          "id",
          "accommodation_id",
          "score",
          "author",
          "created_at"
        ]
      },
      "ReviewsResponse": {
        "type": "object",
        "properties": {
          "next_page": {
            "type": "string",
            "nullable": true
          },
          "request_id": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Resena"
            }
          }
        },
        "required": [
          "next_page",
          "request_id",
          "data"
        ]
      },
...
      "ImagenDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "URL https de la imagen",
            "maxLength": 500
          },
          "es_portada": {
            "type": "boolean"
          }
        },
        "required": [
          "url"
        ]
      },
      "AeropuertoCercanoDto": {
        "type": "object",
        "properties": {
          "aeropuerto_id": {
            "type": "number",
            "minimum": 1
          },
          "distancia_km": {
            "type": "number",
            "minimum": 0,
            "maximum": 9999
          },
          "tiempo_min": {
            "type": "number",
            "minimum": 0,
            "maximum": 3000
          },
          "ofrece_transfer": {
            "type": "boolean"
          }
        },
        "required": [
          "aeropuerto_id",
          "distancia_km",
          "tiempo_min"
        ]
      },
      "CrearAlojamientoDto": {
        "type": "object",
        "properties": {
          "tipo_id": {
            "type": "number",
            "minimum": 1
          },
          "ciudad_id": {
            "type": "number",
            "minimum": 1
          },
          "politica_id": {
            "type": "number",
            "minimum": 1
          },
          "nombre": {
            "type": "string",
            "minLength": 3,
            "maxLength": 140
          },
          "descripcion": {
            "type": "string",
            "minLength": 20,
            "maxLength": 5000
          },
          "direccion": {
            "type": "string",
            "minLength": 5,
            "maxLength": 200
          },
          "latitud": {
            "type": "number",
            "description": "Latitud dentro de Ecuador (-5 a 2)",
            "minimum": -5,
            "maximum": 2
          },
          "longitud": {
            "type": "number",
            "description": "Longitud dentro de Ecuador (-93 a -75, incluye Galápagos)",
            "minimum": -93,
            "maximum": -75
          },
          "telefono_contacto": {
            "type": "string",
            "pattern": "TELEFONO_EC"
          },
          "hora_checkin": {
            "type": "string",
            "pattern": "HORA"
          },
          "hora_checkout": {
            "type": "string",
            "pattern": "HORA"
          },
          "noches_min": {
            "type": "number",
            "minimum": 1,
            "maximum": 365
          },
          "noches_max": {
            "type": "number",
            "minimum": 1,
            "maximum": 365
          },
          "tarifa_limpieza": {
            "type": "number",
            "minimum": 0,
            "maximum": 10000
          },
          "modo_reserva": {
            "type": "string",
            "enum": [
              "INSTANTANEA"
            ]
          },
          "reglas_casa": {
            "type": "string",
            "maxLength": 3000
          },
          "categoria_estrellas": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "registro_turismo": {
            "type": "string",
            "maxLength": 30
          },
          "luaf": {
            "type": "string",
            "maxLength": 30
          },
          "amenidades": {
            "description": "Ids de amenidades (reemplaza la lista completa)",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "imagenes": {
            "description": "Imágenes (reemplaza la lista completa)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImagenDto"
            }
          },
          "aeropuertos": {
            "description": "Aeropuertos cercanos (reemplaza la lista completa)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AeropuertoCercanoDto"
            }
          }
        },
        "required": [
          "tipo_id",
          "ciudad_id",
          "politica_id",
          "nombre",
          "descripcion",
          "direccion",
          "latitud",
          "longitud"
        ]
      },
      "ActualizarAlojamientoDto": {
        "type": "object",
        "properties": {
          "tipo_id": {
            "type": "number",
            "minimum": 1
          },
          "ciudad_id": {
            "type": "number",
            "minimum": 1
          },
          "politica_id": {
            "type": "number",
            "minimum": 1
          },
          "nombre": {
            "type": "string",
            "minLength": 3,
            "maxLength": 140
          },
          "descripcion": {
            "type": "string",
            "minLength": 20,
            "maxLength": 5000
          },
          "direccion": {
            "type": "string",
            "minLength": 5,
            "maxLength": 200
          },
          "latitud": {
            "type": "number",
            "description": "Latitud dentro de Ecuador (-5 a 2)",
            "minimum": -5,
            "maximum": 2
          },
          "longitud": {
            "type": "number",
            "description": "Longitud dentro de Ecuador (-93 a -75, incluye Galápagos)",
            "minimum": -93,
            "maximum": -75
          },
          "telefono_contacto": {
            "type": "string",
            "pattern": "TELEFONO_EC"
          },
          "hora_checkin": {
            "type": "string",
            "pattern": "HORA"
          },
          "hora_checkout": {
            "type": "string",
            "pattern": "HORA"
          },
          "noches_min": {
            "type": "number",
            "minimum": 1,
            "maximum": 365
          },
          "noches_max": {
            "type": "number",
            "minimum": 1,
            "maximum": 365
          },
          "tarifa_limpieza": {
            "type": "number",
            "minimum": 0,
            "maximum": 10000
          },
          "modo_reserva": {
            "type": "string",
            "enum": [
              "INSTANTANEA"
            ]
          },
          "reglas_casa": {
            "type": "string",
            "maxLength": 3000
          },
          "categoria_estrellas": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "registro_turismo": {
            "type": "string",
            "maxLength": 30
          },
          "luaf": {
            "type": "string",
            "maxLength": 30
          },
          "amenidades": {
            "description": "Ids de amenidades (reemplaza la lista completa)",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "imagenes": {
            "description": "Imágenes (reemplaza la lista completa)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImagenDto"
            }
          },
          "aeropuertos": {
            "description": "Aeropuertos cercanos (reemplaza la lista completa)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AeropuertoCercanoDto"
            }
          }
        }
      },
      "UnidadDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "capacidad_huespedes": {
            "type": "number",
            "minimum": 1,
            "maximum": 30
          },
          "num_habitaciones": {
            "type": "number",
            "minimum": 0,
            "maximum": 50
          },
          "num_camas": {
            "type": "number",
            "minimum": 1,
            "maximum": 100
          },
          "num_banos": {
            "type": "number",
            "minimum": 0,
            "maximum": 50
          },
          "cantidad": {
            "type": "number",
            "description": "Unidades iguales a la venta (habitaciones de ese tipo)",
            "minimum": 1,
            "maximum": 500
          },
          "precio_noche_base": {
            "type": "number",
            "minimum": 1,
            "maximum": 100000
          },
          "activa": {
            "type": "boolean"
          }
        },
        "required": [
          "nombre",
          "capacidad_huespedes",
          "precio_noche_base"
        ]
      },
      "ActualizarUnidadDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "capacidad_huespedes": {
            "type": "number",
            "minimum": 1,
            "maximum": 30
          },
          "num_habitaciones": {
            "type": "number",
            "minimum": 0,
            "maximum": 50
          },
          "num_camas": {
            "type": "number",
            "minimum": 1,
            "maximum": 100
          },
          "num_banos": {
            "type": "number",
            "minimum": 0,
            "maximum": 50
          },
          "cantidad": {
            "type": "number",
            "description": "Unidades iguales a la venta (habitaciones de ese tipo)",
            "minimum": 1,
            "maximum": 500
          },
          "precio_noche_base": {
            "type": "number",
            "minimum": 1,
            "maximum": 100000
          },
          "activa": {
            "type": "boolean"
          }
        }
      },
      "DiaCalendarioDto": {
        "type": "object",
        "properties": {
          "fecha": {
            "type": "string",
            "pattern": "FECHA"
          },
          "precio_noche": {
            "type": "number",
            "nullable": true,
            "description": "Precio especial; null = precio base",
            "minimum": 1
          },
          "cantidad_a_la_venta": {
            "type": "number",
            "nullable": true,
            "description": "Unidades a la venta ese día; 0 = cerrado; null = cantidad base",
            "minimum": 0
          }
        },
        "required": [
          "fecha"
        ]
      },
      "CalendarioDto": {
        "type": "object",
        "properties": {
          "dias": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiaCalendarioDto"
            }
          }
        },
        "required": [
          "dias"
        ]
      },
      "ResponderSolicitudDto": {
        "type": "object",
        "properties": {
          "acepta": {
            "type": "boolean"
          }
        },
        "required": [
          "acepta"
        ]
      },
      "ResponderResenaDto": {
        "type": "object",
        "properties": {
          "respuesta": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2000
          }
        },
        "required": [
          "respuesta"
        ]
      },
...
    }
  }
}